+++ CHANGELOG.rst
@@ -2,6 +2,12 @@
 Changelog for package rospilot
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+1.0.0 (2015-08-08)
+------------------
+* Switch to H264 codec for streaming and recording
+* Fix a variety of minor bugs
+* Contributors: Christopher Berner
+
 0.2.5 (2015-05-31)
 ------------------
 * Tune MFC encoder parameters